Refactor cxBatch method from the FlowService into SastScannerService class
